+++ src/test/java/lotto/numbers/AllNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.numbers;
 
 import static org.assertj.core.api.Assertions.*;
